FIX: Another patch to stop snapshots being deleted when alternatves are uploaded
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@9867 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/preview_preprocessing.php
@@ -561,13 +561,18 @@ else if (($ffmpeg_fullpath!=false) && !isset($newfile) && in_array($extension, $
     {
     debug('FFMPEG-VIDEO: Start process for creating previews...', RESOURCE_LOG_APPEND_PREVIOUS);
 	
-	//If we are recreating previews, we should remove the previously created snapshots
-	$directory = dirname(get_resource_path($ref,true,"pre",true) );
-	foreach (glob($directory . "/*") as $filetoremove) {
-		if (strpos($filetoremove, 'snapshot_')!==false){
-			unlink($filetoremove);
+	if($alternative==-1)
+		{
+		//If we are recreating previews, we should remove the previously created snapshots
+		$directory = dirname(get_resource_path($ref,true,"pre",true) );
+		foreach (glob($directory . "/*") as $filetoremove)
+			{
+			if (strpos($filetoremove, 'snapshot_')!==false)
+				{
+				unlink($filetoremove);
+				}
+			}
 		}
-    }
 	
     $snapshottime = 1;
